[W1d] bc-hrbl: type worker_count as size_t
bc_hrbl_writer_options_t.worker_count etait uint32_t alors que tous les
autres compteurs internes (bc_concurrency_config_t.worker_count,
bc_hrbl_encoder_pool_t.worker_count, parametres de fonctions) sont en
size_t. Cast (size_t) elimine d'office.

Audit code-review 2026-04-30 sect. B.3, P0 #7.

Pipeline 27 gates verts (asan/tsan/ubsan/memcheck, cppcheck, format,
build/test release+debug, cascade DAG : bc-hrbl-cli, bc-hash,
bc-integrity tous PASS).

Refs: audits/2026-04-30-plan-implementation-bc-libraries.md sect. W1d

+++ include/public/bc_hrbl_writer.h
@@ -11,7 +11,7 @@
 #include <stdint.h>
 
 typedef struct bc_hrbl_writer_options {
-    uint32_t worker_count;
+    size_t worker_count;
     bool deduplicate_strings;
 } bc_hrbl_writer_options_t;
